feat(admin,api): 2513 - post merge
